Add files via upload
Research paper investigating the Susceptible, Infected, and Recovered (SIR) epidemiological model for COVID-19 in the United States, Canada, and India. I created differential equations and used Euler's method and optimization to analyze infectious diseases and learned how to collect reliable data when researching. At the end of the study, I analyzed the curves to learn how population groups play a role in the spread of diseases.